Move Ocelot.Benchmarks.csproj to benchmark/ folder
