+++ Ocelot.slnx
@@ -1,7 +1,7 @@
 <Solution>
-  <Project Path="acceptance/Ocelot.Acceptance.csproj" />
   <Project Path="src/Ocelot.csproj" />
-  <Project Path="test/Ocelot.Benchmarks/Ocelot.Benchmarks.csproj" />
+  <Project Path="acceptance/Ocelot.Acceptance.csproj" />
+  <Project Path="benchmark/Ocelot.Benchmarks.csproj" />
   <Project Path="test/Ocelot.ManualTest/Ocelot.ManualTest.csproj" />
   <Project Path="testing/Ocelot.Testing.csproj" />
   <Project Path="unit/Ocelot.UnitTests.csproj" />
