+++ benchmark/MsLoggerBenchmarks.cs
