+++ benchmark/SerilogBenchmarks.cs
