--- test/Ocelot.Benchmarks/Properties/AssemblyInfo.cs
@@ -1,16 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following set of attributes.
-// Change these attribute values to modify the information associated with an assembly.
-[assembly: AssemblyCompany("Three Mammals")]
-[assembly: AssemblyCopyright("© 2026 Three Mammals. MIT licensed OSS.")]
-[assembly: AssemblyProduct("Ocelot Gateway")]
-[assembly: AssemblyTrademark("Ocelot")]
-
-// Setting ComVisible to false makes the types in this assembly not visible to COM components.
-// If you need to access a type in this assembly from COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("106b49e6-95f6-4a7b-b81c-96bfa74af035")]
